+++ packages/kysely/src/migrate/migrations/2024-12-05T13:05:00_drop_tickets.ts
@@ -1,0 +1,29 @@
+import { Kysely } from "kysely";
+import { Database } from "@authhero/kysely-adapter";
+
+export async function up(db: Kysely<Database>): Promise<void> {
+  await db.schema.dropTable("tickets").execute();
+}
+
+export async function down(db: Kysely<Database>): Promise<void> {
+  await db.schema
+    .createTable("tickets")
+    .addColumn("tenant_id", "varchar(255)", (col) =>
+      col.references("tenants.id").onDelete("cascade").notNull(),
+    )
+    .addColumn("id", "varchar(255)", (col) => col.primaryKey())
+    .addColumn("client_id", "varchar(255)", (col) =>
+      col.references("applications.id").onDelete("cascade").notNull(),
+    )
+    .addColumn("email", "varchar(255)", (col) => col.notNull())
+    .addColumn("nonce", "varchar(255)")
+    .addColumn("state", "varchar(1024)")
+    .addColumn("scope", "varchar(1024)")
+    .addColumn("response_type", "varchar(256)")
+    .addColumn("response_mode", "varchar(256)")
+    .addColumn("redirect_uri", "varchar(1024)")
+    .addColumn("created_at", "varchar(255)", (col) => col.notNull())
+    .addColumn("expires_at", "varchar(255)", (col) => col.notNull())
+    .addColumn("used_at", "varchar(255)")
+    .execute();
+}